refactor: ChainFacadeI
